+++ .storybook/main.ts
@@ -1,4 +1,5 @@
 import path from "path"
+import webpack from "webpack"
 
 module.exports = {
   "stories": [
@@ -16,6 +17,14 @@ module.exports = {
     config.resolve.alias['swr'] = require.resolve("./swr-mock.tsx");
     // mock our env vars so they can be updated via the toolbar
     config.resolve.alias['utils/env'] = require.resolve("./env-mock.ts");
+
+    // mock out the decryptor like we do in next.config.js
+    config.plugins.push(
+      new webpack.NormalModuleReplacementPlugin(
+        /@nypl-simplified-packages\/axisnow-access-control-web/,
+        "utils/mockDecryptor.ts"
+      )
+    );
 
     config.resolve.modules = [
       path.resolve(__dirname, "../src"),