+++ __mocks__/@nypl-simplified-packages/axisnow-access-control-web/index.ts
@@ -1,0 +1,11 @@
+export class AxisNowDecryptionError {
+  constructor(...args: any[]) {
+    console.log("Instantiated an AxisNowDecryptionError with args: ", ...args);
+  }
+}
+
+export default class Decryptor {
+  constructor(...args: any[]) {
+    console.log("Instantiated an AxisNowDecryptor with args: ", ...args);
+  }
+}