+++ src/utils/mockDecryptor.ts
@@ -1,0 +1,9 @@
+/**
+ * This file is to mock out @nypl-simplified-packages/axisnow-access-control-web
+ * so that builds that don't have access to it can still succeed. It is not meant
+ * to provide any functionality, and relying on the mock will result in an exception.
+ */
+
+const Decryptor = undefined;
+export default Decryptor;
+export const AxisNowDecryptionError = undefined;